requirements.txt
flask
elasticsearch
requests
slack_sdk
reportlab
pdfkit
python-dotenv 